% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LEEA.R
\name{run_LEEA}
\alias{run_LEEA}
\title{Run Loop Eigenvalue Elasticity Analysis}
\usage{
run_LEEA(times, graph, sim_df, method = "analytical")
}
\arguments{
\item{times}{A numeric vector that indicates in which time steps the
procedure is applied.}

\item{graph}{An igraph object representation of a System Dynamics model.}

\item{sim_df}{A dataframe with the simulation results of a System Dynamics
model. This dataframe must contain the column 'time'.}

\item{method}{A string indicating the method for estimating edge gains. It
can be either 'analytical' or 'numerical'.}
}
\description{
Run Loop Eigenvalue Elasticity Analysis
}
