% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{struc_eval_impact}
\alias{struc_eval_impact}
\title{Evaluate impact equations numerically}
\usage{
struc_eval_impact(impact_df, sim_df)
}
\arguments{
\item{impact_df}{Data frame returned by the function \code{struc_impact_on}}

\item{sim_df}{Data frame with the simulation results}
}
\value{

}
\description{
Evaluate impact equations numerically
}
\examples{
I_z_x <- "-(R * x) * (a * y - z) / (-R * x * z)"
I_x_x <- "-(R * z) * (-R * x * z) / (-R * x * z)"

impact_df <- data.frame(from    = c("z", "x"),
                        to      = c("x", "x"),
                        through = c("f1", "f1"),
                        impact  = c(I_z_x, I_x_x))

x0  <- 0.974182771
x10 <- 0.466656479
y0  <- 0.009831036
y10 <- 0.078797261
z0  <- 0.007933189
z10 <- 0.078490487

sim_df <- data.frame(
  time = c(0L, 10L),
  x = c(x0, x10),
  y = c(y0, y10),
  z = c(z0, z10),
  R = c(2L, 2L),
  a = c(1L, 1L))
struc_eval_impact(impact_df, sim_df)
}
